Create DailyAdviceServer
This program makes a ServerSocket and waits for client requests. When it
gets a client request (i.e. client said new Socket() for this
application), the server makes a new Socket connection to that client.
The server makes a PrintWriter using the Socket's output stream) and
sends a message to the client.